Re-extract plgg-mcp as the content-free MCP protocol substrate
Why: The ticket measured that the MCP surface drags plgg-content and node:sqlite into every consumer (fatal on bun); the developer chose re-extraction over subpath exports: a substrate that reaches a SQLite store is not a substrate, and ticket 27's HTTP+OAuth transport needs the clean seam.

Changes: New packages/plgg-mcp@0.0.2 (0.0.1 is frozen on npm) carries Mcp/Rpc/Transport/vendors moved via git mv from plgg-cms's internal mcpProtocol, depending only on plgg; plgg-cms@0.0.3 keeps the content Tools and re-exports both through its unchanged mcpProtocol barrel; build/install scripts and READMEs wired; a moduleGraph pin spec asserts node:sqlite, plgg-sql, plgg-cms, and plgg-content absent from the package source and the dependency set exactly plgg.

Concerns: The frozen npm consumer on plgg-mcp@0.0.1 still needs its own migration decision - 0.0.2 is a different (smaller) surface: contentTools moved to plgg-cms. Publishing both packages is the developer-driven gate at ship.

Insights: The pin is the durable part: the module-graph spec restates the unrunnable bun gate as a source-level assertion that fails on the first convenience re-export dragging the store back in. git mv preserved history across the package boundary.

Verify: plgg-mcp: tsc clean, 16 tests (incl. 3 pin tests), coverage gate passed, dist emits zero node:sqlite references; plgg-cms: tsc clean, 492 tests, coverage gate passed, dist imports plgg-mcp as external; gate-readme and vendor-boundary green; fresh check-all EXIT 0

Category: Changed
