+++ README.md
@@ -106,7 +106,8 @@ its name); this section is the top-level index that links down to each.
 - **[`packages/plgg-highlight/`](packages/plgg-highlight/)** - Zero-dep TS/TSX/JS/JSX/JSON syntax highlighting for plgg-md's `Highlighter` seam, tokenizing with an in-house plgg-parser grammar (no `typescript` dependency) into classified plgg-view `Html` spans
 - **[`packages/plgg-parser/`](packages/plgg-parser/)** - Zero-new-dep generic parser combinator library on plgg: data-last `Parser<A,S>` functions returning `Result<Parsed<A,S>>`, a user-state slot for context-sensitive grammars, and a TS-lexer demo (the eventual in-house replacement for plgg-highlight's compiler scanner)
 - **[`packages/plggpress/`](packages/plggpress/)** - The slim VitePress-like static-site generator on the plgg family: a typed `SiteConfig` contract, a base-path href resolver, a config/`build` CLI, and a build-time dead-link checker — the engine that builds the guide. Carries its generic web-application framework internally and exposes it as a public `plggpress/framework` subpath (config loading, a router builder, static-build + CLI orchestration). No CMS/server dependencies — the dynamic content surface lives in `plgg-cms`
-- **[`packages/plgg-cms/`](packages/plgg-cms/)** - The dynamic content-management surface that pairs with plggpress: CMS-owned content indexing/query, admin UI, durable-domain derivation, a read-only content delivery API, OIDC auth, content editing, media, stakeholder submission, ops, MCP protocol/tools, and agent surfaces — composed onto plggpress's `framework` seam and served as an always-on `node:http` instance (the `plgg-cms serve` bin, D5's dynamic half)
+- **[`packages/plgg-cms/`](packages/plgg-cms/)** - The dynamic content-management surface that pairs with plggpress: CMS-owned content indexing/query, admin UI, durable-domain derivation, a read-only content delivery API, OIDC auth, content editing, media, stakeholder submission, ops, the content MCP tools, and agent surfaces — composed onto plggpress's `framework` seam and served as an always-on `node:http` instance (the `plgg-cms serve` bin, D5's dynamic half)
+- **[`packages/plgg-mcp/`](packages/plgg-mcp/)** - The MCP protocol substrate: typed tool registry + dispatch (`initialize`, `tools/list`, `tools/call`), JSON-RPC 2.0 framing, transport-agnostic frame handling, and a stdio server seam — no content store, no default tool set, `node:sqlite` pinned absent from its module graph; plgg-cms composes its content tools onto it
 - **[`packages/plgg-bundle/`](packages/plgg-bundle/)** - In-house minimal library bundler (dual ESM+CJS output + a per-file `.d.ts` tree) and dev server, plgg-free with zero new dependencies (reuses the project's own TypeScript)
 - **[`packages/plgg-test/`](packages/plgg-test/)** - In-house minimal test runner (the `plgg-test` bin every package's test/coverage scripts call): discovery, assertions/matchers, mocks, and a coverage threshold gate
