--- packages/plgg-cms/src/mcpProtocol/Mcp/index.ts
@@ -1,2 +1,0 @@
-export * from "plgg-cms/mcpProtocol/Mcp/model";
-export * from "plgg-cms/mcpProtocol/Mcp/usecase";
