--- packages/plgg-cms/src/mcpProtocol/Mcp/model/index.ts
@@ -1,1 +1,0 @@
-export * from "plgg-cms/mcpProtocol/Mcp/model/Tool";
