--- packages/plgg-cms/src/mcpProtocol/Mcp/usecase/index.ts
@@ -1,1 +1,0 @@
-export * from "plgg-cms/mcpProtocol/Mcp/usecase/dispatchMcp";
