--- packages/plgg-cms/src/mcpProtocol/Rpc/index.ts
@@ -1,2 +1,0 @@
-export * from "plgg-cms/mcpProtocol/Rpc/model";
-export * from "plgg-cms/mcpProtocol/Rpc/usecase";
