--- packages/plgg-cms/src/mcpProtocol/Rpc/model/index.ts
@@ -1,1 +1,0 @@
-export * from "plgg-cms/mcpProtocol/Rpc/model/RpcMessage";
