--- packages/plgg-cms/src/mcpProtocol/Rpc/usecase/index.ts
@@ -1,2 +1,0 @@
-export * from "plgg-cms/mcpProtocol/Rpc/usecase/parseRequest";
-export * from "plgg-cms/mcpProtocol/Rpc/usecase/serializeResponse";
