--- packages/plgg-cms/src/mcpProtocol/Transport/index.ts
@@ -1,1 +1,0 @@
-export * from "plgg-cms/mcpProtocol/Transport/handleFrame";
