--- packages/plgg-cms/src/mcpProtocol/vendors/index.ts
@@ -1,1 +1,0 @@
-export * from "plgg-cms/mcpProtocol/vendors/stdioServer";
