+++ packages/plgg-mcp/.prettierrc.json
@@ -1,0 +1,7 @@
+{
+  "printWidth": 50,
+  "semi": true,
+  "singleQuote": false,
+  "trailingComma": "all",
+  "bracketSameLine": false
+}
