+++ packages/plgg-mcp/plgg-test.config.json
@@ -1,0 +1,9 @@
+{
+  "coverage": {
+    "threshold": 90,
+    "exclude": [
+      "/index.ts",
+      "/vendors/stdioServer.ts"
+    ]
+  }
+}
