+++ packages/plgg-mcp/src/Mcp/model/index.ts
@@ -1,0 +1,1 @@
+export * from "plgg-mcp/Mcp/model/Tool";
