+++ packages/plgg-mcp/src/Mcp/usecase/index.ts
@@ -1,0 +1,1 @@
+export * from "plgg-mcp/Mcp/usecase/dispatchMcp";
