+++ packages/plgg-mcp/src/Rpc/model/index.ts
@@ -1,0 +1,1 @@
+export * from "plgg-mcp/Rpc/model/RpcMessage";
