+++ packages/plgg-mcp/src/vendors/index.ts
@@ -1,0 +1,1 @@
+export * from "plgg-mcp/vendors/stdioServer";
