+++ packages/plgg-mcp/tsconfig.json
@@ -1,0 +1,38 @@
+{
+  "compilerOptions": {
+    /* Compiler Option */
+    "target": "ES2021",
+    "module": "NodeNext",
+    "lib": [
+      "ES2021"
+    ],
+    "types": [
+      "node"
+    ],
+    "rootDir": "src",
+    "noEmit": true,
+    "paths": {
+      "plgg-mcp*": [
+        "./src/*"
+      ]
+    },
+    /* Type Checking Rules */
+    "strict": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "noUncheckedIndexedAccess": true,
+    "noImplicitReturns": true,
+    "noFallthroughCasesInSwitch": true,
+    "allowUnusedLabels": false,
+    "allowUnreachableCode": false,
+    "forceConsistentCasingInFileNames": true,
+    "exactOptionalPropertyTypes": true,
+    "skipLibCheck": true,
+    "isolatedModules": true,
+    "allowJs": false,
+    "erasableSyntaxOnly": true
+  },
+  "include": [
+    "src/**/*"
+  ]
+}
